+++ src/collective/documentgenerator/setuphandlers.py
@@ -120,7 +120,7 @@ def install_demo(context):
             ),
             container=pod_folder,
             excludeFromNav=True,
-            pod_formats=['odt', 'pdf', ],
+            pod_formats=['odt', 'pdf', 'doc', 'docx'],
             pod_portal_types=['Document'],
             style_template=[style_template.UID()],
             merge_templates=[